backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Ahmed Gamal"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.3"
djangorestframework = "^3.14.0"
markdown = "^3.5.2"
django-filter = "^23.5"
pandas = "^2.2.0"
django-cors-headers = "^4.3.1"
psycopg2-binary = "^2.9.9"
numpy = "^1.26.3"
uvicorn = "^0.27.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
